+++ src/together/types/fine_tuning_list_response.py
@@ -208,6 +208,9 @@ class Data(BaseModel):
     wandb_name: Optional[str] = None
     """Weights & Biases run name"""
 
+    wandb_entity: Optional[str] = None
+    """Weights & Biases entity name"""
+
     wandb_project_name: Optional[str] = None
     """Weights & Biases project name"""